+++ src/renderer/components/Composer.js
@@ -1,30 +1,23 @@
-import React, { useRef, useState, useEffect } from 'react'
-import { withTheme } from 'styled-components'
+import React, { useRef, useState } from 'react'
 import { Button } from '@blueprintjs/core'
 import { remote } from 'electron'
-import styled from 'styled-components'
-import { Picker } from 'emoji-mart'
-import classNames from 'classnames'
 
 import SettingsContext from '../contexts/SettingsContext'
 import ComposerMessageInput from './ComposerMessageInput'
-import { callDcMethodAsync, callDcMethod } from '../ipc'
 import logger from '../../logger'
+import EmojiAndStickerPicker from './EmojiAndStickerPicker'
 
 const log = logger.getLogger('renderer/composer')
 
-const Composer = withTheme(React.forwardRef((props, ref) => {
-  const { onSubmit, isDisabled, chatId, draft, theme } = props
+const Composer = React.forwardRef((props, ref) => {
+  const { onSubmit, isDisabled, chatId, draft } = props
 
   const [filename, setFilename] = useState(null)
-  const [error, setError] = useState(false)
   const [showEmojiPicker, setShowEmojiPicker] = useState(false)
 
   const messageInputRef = useRef()
   const pickerRef = useRef()
   const pickerButtonRef = useRef()
-
-  const handleError = () => setError(true)
 
   const sendMessage = () => {
     const message = messageInputRef.current.getText()
@@ -32,14 +25,15 @@ const Composer = withTheme(React.forwardRef((props, ref) => {
       log.debug(`Empty message: don't send it...`)
       return
     }
-    onSubmit({text: message})
+    onSubmit({ text: message, filename })
+
     messageInputRef.current.clearText()
     messageInputRef.current.focus()
   }
 
   const addFilename = () => {
     remote.dialog.showOpenDialog({ properties: ['openFile'] }, filenames => {
-      if (filenames && filenames[0]) onSubmit({ filename: filenames[0] })
+      if (filenames && filenames[0]) setFilename(filenames[0])
     })
   }
 
@@ -50,12 +44,12 @@ const Composer = withTheme(React.forwardRef((props, ref) => {
     messageInputRef.current.insertStringAtCursorPosition(emoji.native)
   }
 
-
   const tx = window.translate
 
   if (isDisabled) {
-    return <div className='composer' ref={ref}
-      className={'disabled-message-input'}
+    return <div
+      ref={ref}
+      className={'composer composer--disabled-message-input'}
       style={{ textAlign: 'center', padding: '0.5rem', color: '#999' }}>
       {tx('messaging_disabled_not_in_group')}
     </div>
@@ -84,7 +78,7 @@ const Composer = withTheme(React.forwardRef((props, ref) => {
           <span />
         </div>
         { showEmojiPicker &&
-            <EmojiAndStickerPicker chatId={chatId} ref={pickerRef} onEmojiSelect={onEmojiSelect} color={theme.emojiSelectorSelectionColor} />
+        <EmojiAndStickerPicker chatId={chatId} ref={pickerRef} onEmojiSelect={onEmojiSelect} />
         }
         <div className='composer__send-button-wrapper' onClick={sendMessage}>
           <button aria-label={tx('a11y_send_btn_label')} />
@@ -92,103 +86,6 @@ const Composer = withTheme(React.forwardRef((props, ref) => {
       </div>
     )
   }
-}))
+})
 
 export default Composer
-
-export const useAsyncEffect = (asyncEffect, ...args) => useEffect(() => { asyncEffect() }, ...args)
-
-
-export const StickerDiv = props => {
-  const { stickerPackName, stickerPackImages, chatId } = props
-  const onClickSticker = (fileName) => {
-    callDcMethod('sendMessage', [chatId, 'We are testing stickers', fileName])
-  }
-
-  return (
-    <div>
-      <div className='emoji-sticker-picker__sticker-picker__inner__sticker-pack-title'>{stickerPackName}</div>
-      <div className='emoji-sticker-picker__sticker-picker__inner__sticker-pack-container'>
-        { stickerPackImages.map((filePath, index) => {
-            return (
-              <div className='emoji-sticker-picker__sticker-picker__inner__sticker-pack-container__sticker'>
-                <img
-                  key={index}
-                  src={filePath}
-                  onClick={onClickSticker.bind(this, filePath)}
-                />
-              </div>
-            )
-        })}
-      </div>
-    </div>
-  )
-}
-
-export const StickerPicker = props => {
-  const { stickers, chatId } = props
-  return (
-    <div className='emoji-sticker-picker__sticker-picker'>
-      <div className='emoji-sticker-picker__sticker-picker__inner'>
-       { Object.keys(stickers).map(stickerPackName => {
-           return <StickerDiv
-             chatId={chatId}
-             key={stickerPackName}
-             stickerPackName={stickerPackName}
-             stickerPackImages={stickers[stickerPackName]}
-           />
-       })}
-      </div>
-    </div>
-  )
-}
-
-export const EmojiOrStickerSelectorButton = props => {
-  return (
-    <div
-      className={classNames(
-        'emoji-sticker-picker__emoji-or-sticker-selector__button',
-        { 'emoji-sticker-picker__emoji-or-sticker-selector__button--is-selected' : props.isSelected })}
-      onClick={props.onClick}
-    >
-      {props.children}
-    </div>
-  )
-}
-
-export const EmojiAndStickerPicker = React.forwardRef((props, ref) => {
-  const { color, onEmojiSelect, emojiTooltip, chatId } = props
-
-  const [showSticker, setShowSticker] = useState(false)
-  const [stickers, setStickers] = useState(null)
-  
-  useAsyncEffect(async () => {
-    const stickers = await callDcMethodAsync('getStickers')
-    setStickers(stickers)
-  }, [])
-
-  return (
-    <div className='emoji-sticker-picker'>
-      <div className='emoji-sticker-picker__emoji-or-sticker-selector'>
-        <EmojiOrStickerSelectorButton onClick={() => setShowSticker(false)} isSelected={!showSticker}>Emoji</EmojiOrStickerSelectorButton>
-        <EmojiOrStickerSelectorButton onClick={() => setShowSticker(true)} isSelected={showSticker}>Sticker</EmojiOrStickerSelectorButton>
-      </div>
-      <div className='emoji-sticker-picker__emoji-or-sticker-picker'>
-       { !showSticker &&
-         <div className='emoji-sticker-picker__emoji-picker' ref={ref}>
-           <Picker
-             style={{ width: '100%', height: '100%' }}
-             native
-             color={color}
-             onSelect={onEmojiSelect}
-             showPreview={false}
-             showSkinTones={false}
-             emojiTooltip
-           />
-         </div>
-       }
-       { showSticker && stickers !== null && typeof stickers === 'object' && <StickerPicker chatId={chatId} stickers={stickers} /> }
-      </div>
-    </div>
-  )
-})
